% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrettext_funs.R
\name{decode}
\alias{decode}
\title{Decrypt a Message}
\usage{
decode(text, key1, key2)
}
\arguments{
\item{text}{A string}

\item{key1}{A numeric value between 1 and 25}

\item{key2}{A numeric value between 1 and 25}
}
\value{
A string, converted to lowercase and decrypted
}
\description{
Decrypt a character string generated with setcode() given two numeric keys.
}
\examples{
# string argument as output of setcode()
decode(setcode("hello world", 5, 16), 5, 16)
# string argument as user defined character string
decode("cjakx preik", 5, 16)
}
