% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrettext_funs.R
\name{setcode}
\alias{setcode}
\title{Encrypt a message}
\usage{
setcode(text, key1, key2)
}
\arguments{
\item{text}{A string}

\item{key1}{A numeric value between 1 and 25}

\item{key2}{A numeric value between 1 and 25}
}
\value{
A string, converted to lowercase and encrypted
}
\description{
Encrypt a character string given two numeric keys.
}
\examples{
setcode("hello world", 5, 16)
}
